+++ changelog.md
@@ -3,16 +3,21 @@
 ### Changes
 * Added global localized errorhandling
 * Added IViewLocalizer injection in _ViewImports.cshtml so there's a Localizer available for every view.
+* Changed SupportedCultures from List<string> to HashSet<string>
 
 * Documentation:
     * Renamed LocalController to ExampleController to make it more clear what it's for.
     * Added usages of IViewLocalizer Localizer for the example pages.
+
+* Tests:
+    * TODO!
 
 ### Bugfixes
 
 ### Issues closed
 
 * #5    - [Global errorhandling](../../issues/5)
+* #9    - [Change SupportedCultures to use Dictionary instead of List](../../issues/9)
 * #13   - [Rename LocalController to ExampleController](../../issues/13)
 * #14   - [Add example of IViewLocalizer to Example views](../../issues/14)